+++ packages/side-effects/side-effect-title/types/index.d.ts
@@ -1,7 +1,7 @@
-import { Subscriber } from '@curi/core';
+import { ResponseHandler } from '@curi/core';
 export interface TitleOptions {
     prefix?: string;
     suffix?: string;
 }
-declare function createTitleSideEffect(options?: TitleOptions): Subscriber;
+declare function createTitleSideEffect(options?: TitleOptions): ResponseHandler;
 export default createTitleSideEffect;